Add `AssociativeArray` ability to deallocate keys
`AssociativeArray` by design does not manage keys, automatically
deallocating them, like it can with stored values. This patch adds a
parameter to `Empty()` method that allows it to deallocate all of the
used keys when deallocation all of the values.

As a side effect this patch also fixes a bug that broke
`AssociativeArray` after using `Empty()` method.